# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Online API
  version: 1.0.0
  description: >-
    Kontrak API katalog, reservasi, checkout, pembayaran, e-ticket,
    autentikasi petugas, pengelolaan akun STAFF, dan check-in. Token sesi
    petugas terpisah dari token akses order. Respons error memakai bentuk
    {error: {code, message}}; beberapa respons check-in menyertakan detail tambahan.
    TRANSACTIONS_ENABLED=false menghentikan pembuatan reservasi, checkout, dan pembayaran baru
    dengan HTTP 503 dan kode TRANSACTIONS_PAUSED; webhook, rekonsiliasi, serta akses tiket tetap tersedia.
servers:
  - url: /
    description: Origin aplikasi
...
paths:
  /api/v1/events:
    get:
      operationId: listEvents
      tags: [Katalog]
      summary: Daftar konser
      responses:
        '200':
          description: Daftar konser; array dapat kosong.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
              examples:
                concerts:
                  value:
                    events:
                      - id: nusa-malam
                        artist: Nusa Malam
                        city: Jakarta
                        venue: Ruang Selatan
                        address: Jl. Musik Raya, Jakarta
                        startsAt: '2027-08-24T19:30:00Z'
                        genre: Indie
                        status: Early Bird
                        publicationStatus: PUBLISHED
                        image: https://example.test/nusa-malam.jpg
                        description: Konser Nusa Malam.
                        lineup: [Nusa Malam, Pembuka Sore]
                        price: 225000
                        zones:
                          - {id: festival, name: Festival, description: Area berdiri umum}
                        ticketTiers:
                          - {id: festival, name: Festival, zoneId: festival, price: 225000, availableQuantity: 42, maxPerOrder: 6, benefit: Area berdiri umum, gate: Gate B, seating: free-standing}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events/{eventID}:
    parameters:
      - $ref: '#/components/parameters/EventID'
    get:
      operationId: getEvent
      tags: [Katalog]
      summary: Detail konser dan tier tiket
      responses:
        '200':
          description: Detail konser.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              example:
                id: nusa-malam
                artist: Nusa Malam
                city: Jakarta
                venue: Ruang Selatan
                address: Jl. Musik Raya, Jakarta
                startsAt: '2027-08-24T19:30:00Z'
                genre: Indie
                status: Early Bird
                publicationStatus: PUBLISHED
                image: https://example.test/nusa-malam.jpg
                description: Konser Nusa Malam.
                lineup: [Nusa Malam, Pembuka Sore]
                price: 225000
                zones:
                  - {id: festival, name: Festival, description: Area berdiri umum}
                ticketTiers:
                  - {id: festival, name: Festival, zoneId: festival, price: 225000, availableQuantity: 42, maxPerOrder: 6, benefit: Area berdiri umum, gate: Gate B, seating: free-standing}
        '404':
          description: Konser tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: EVENT_NOT_FOUND, message: Event tidak ditemukan}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reservations/{reservationID}/event:
    get:
      operationId: getReservedEvent
      tags: [Reservasi]
      summary: Memuat event dari reservasi yang sudah dibuat
      description: >-
        Untuk memulihkan checkout reservasi yang masih tersimpan setelah event
        disembunyikan dari katalog. Membutuhkan ID reservasi dan Idempotency-Key.
      parameters:
        - name: reservationID
          in: path
          required: true
          schema: {type: string, minLength: 32, maxLength: 32}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Informasi konser terkait reservasi.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Event'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/reservations:
    post:
      operationId: createReservation
      tags: [Reservasi]
      summary: Tahan stok tiket untuk sementara
      description: >-
        Kirim key yang sama dan payload yang sama untuk mengulang request dengan aman.
        Key harus 16–100 byte setelah trim. Reservation berlaku selama TTL yang
        dikonfigurasi server.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
            example:
              eventId: nusa-malam
              items:
                - {tierId: festival, quantity: 2}
                - {tierId: vip-a, quantity: 1}
      responses:
        '201':
          description: Reservasi dibuat. Replay key yang sama juga mengembalikan 201.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
              example:
                id: 0123456789abcdef0123456789abcdef
                status: ACTIVE
                expiresAt: '2027-08-24T12:40:00Z'
                event: {id: nusa-malam, artist: Nusa Malam}
                items:
                  - {tierId: festival, name: Festival, quantity: 2, unitPrice: 225000, lineTotal: 450000}
                subtotal: 450000
        '400':
          description: Key hilang/tidak valid, JSON tidak valid, atau data reservasi tidak valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidKey:
                  value: {error: {code: INVALID_REQUEST, message: 'Idempotency-Key tidak valid'}}
                invalidJSON:
                  value: {error: {code: INVALID_JSON, message: 'JSON request tidak valid'}}
                invalidRequest:
                  value: {error: {code: INVALID_REQUEST, message: 'Data reservasi tidak valid'}}
        '404':
          description: Event atau tier tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                event:
                  value: {error: {code: EVENT_NOT_FOUND, message: Event tidak ditemukan}}
                tier:
                  value: {error: {code: TIER_NOT_FOUND, message: Tier tiket tidak ditemukan}}
        '409':
          description: Stok kurang atau key telah digunakan untuk payload berbeda.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                stock:
                  value: {error: {code: INSUFFICIENT_STOCK, message: Stok tiket tidak mencukupi}}
                idempotency:
                  value: {error: {code: IDEMPOTENCY_CONFLICT, message: Idempotency-Key sudah digunakan untuk request berbeda}}
        '422':
          description: Jumlah melampaui batas pemesanan tier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: ORDER_LIMIT_EXCEEDED, message: Jumlah tiket melebihi batas pemesanan}
        '503':
          description: Pembuatan transaksi baru sedang dihentikan.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reservations/{reservationID}:
    parameters:
      - $ref: '#/components/parameters/ReservationID'
    get:
      operationId: getReservation
      tags: [Reservasi]
      summary: Baca status reservasi
      responses:
        '200':
          description: Reservasi dan snapshot item yang ditahan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservasi tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_NOT_FOUND, message: Reservasi tidak ditemukan}
        '410':
          description: Reservasi kedaluwarsa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_EXPIRED, message: Reservasi sudah kedaluwarsa}
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: cancelReservation
      tags: [Reservasi]
      summary: Batalkan reservasi aktif dan kembalikan stok
      responses:
        '200':
          description: Reservasi yang dibatalkan; pembatalan ulang mengembalikan status yang sama.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservasi tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_NOT_FOUND, message: Reservasi tidak ditemukan}
        '409':
          description: Reservasi telah dikonversi atau dibatalkan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                converted:
                  value: {error: {code: RESERVATION_CONVERTED, message: Reservasi sudah dikonversi}}
                cancelled:
                  value: {error: {code: RESERVATION_CANCELLED, message: Reservasi sudah dibatalkan}}
        '410':
          description: Reservasi kedaluwarsa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_EXPIRED, message: Reservasi sudah kedaluwarsa}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reservations/{reservationID}/convert:
    parameters:
      - $ref: '#/components/parameters/ReservationID'
    post:
      operationId: convertReservation
      tags: [Reservasi]
      summary: Konversi reservasi ke tiket demo browser
      description: >-
        Endpoint kompatibilitas untuk alur demo saat ini. Endpoint ini mengubah
        reservasi menjadi CONVERTED dan tidak membuat order atau e-ticket database.
        Alur pembelian backend memakai endpoint checkout di bawah.
      responses:
        '200':
          description: Reservasi yang sudah dikonversi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservasi tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_NOT_FOUND, message: Reservasi tidak ditemukan}
        '409':
          description: Reservasi telah dibatalkan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_CANCELLED, message: Reservasi sudah dibatalkan}
        '410':
          description: Reservasi kedaluwarsa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_EXPIRED, message: Reservasi sudah kedaluwarsa}
        '503':
          description: Pembuatan transaksi baru sedang dihentikan.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reservations/{reservationID}/checkout:
    parameters:
      - $ref: '#/components/parameters/ReservationID'
    post:
      operationId: createOrder
      tags: [Checkout]
      summary: Buat order idempoten dari reservasi
      description: >-
        Gunakan Idempotency-Key yang sama dengan saat membuat reservasi. Replay
        request checkout yang sama mengembalikan order dan token akses yang sama.
        Harga, biaya admin, dan diskon dihitung server. Order PENDING kedaluwarsa
        pada expiresAt; tanggal kedaluwarsa token akses berbeda.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              buyer: {name: Nama Pembeli, email: buyer@example.com, phone: '+6281234567890', identity: '123456789012'}
              attendees:
                - {tierId: festival, names: [Peserta Satu, Peserta Dua]}
              voucherCode: HEMAT10
      responses:
        '201':
          description: Order baru dibuat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '200':
          description: Replay dengan payload checkout yang sama; mengembalikan order yang sama.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Key atau ID reservasi tidak valid, atau JSON tidak valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidRequest:
                  value: {error: {code: INVALID_REQUEST, message: Idempotency-Key tidak valid}}
                invalidJSON:
                  value: {error: {code: INVALID_JSON, message: JSON request tidak valid}}
        '401':
          description: Key tidak cocok dengan reservasi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_ACCESS_DENIED, message: Akses reservasi tidak valid}
        '404':
          description: Reservasi atau target order tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                reservation:
                  value: {error: {code: RESERVATION_NOT_FOUND, message: Reservasi tidak ditemukan}}
                order:
                  value: {error: {code: ORDER_NOT_FOUND, message: Order tidak ditemukan}}
        '409':
          description: Reservasi dibatalkan/dikonversi atau data checkout berbeda dari order yang ada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cancelled:
                  value: {error: {code: RESERVATION_CANCELLED, message: Reservasi sudah dibatalkan}}
                converted:
                  value: {error: {code: RESERVATION_CONVERTED, message: Reservasi sudah dikonversi}}
                conflict:
                  value: {error: {code: CHECKOUT_CONFLICT, message: Reservasi sudah dipesan dengan data checkout berbeda}}
        '410':
          description: Reservasi kedaluwarsa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: RESERVATION_EXPIRED, message: Reservasi sudah kedaluwarsa}
        '422':
          description: Data checkout tidak valid atau voucher tidak didukung.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidRequest:
                  value: {error: {code: INVALID_REQUEST, message: Data checkout tidak valid}}
                invalidVoucher:
                  value: {error: {code: INVALID_VOUCHER, message: Kode voucher tidak valid}}
        '503':
          description: Pembuatan transaksi baru sedang dihentikan.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderID}:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    get:
      operationId: getOrder
      tags: [Checkout]
      summary: Baca detail order dan pembayaran
      security:
        - OrderAccessToken: []
      responses:
        '200':
          description: Order, data pembeli, peserta, dan ringkasan pembayaran jika sudah ada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '400':
          $ref: '#/components/responses/InvalidID'
        '401':
          $ref: '#/components/responses/AccessTokenRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/AccessTokenExpired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderID}/payments:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      operationId: createSnapPayment
      tags: [Pembayaran]
      summary: Buat atau lanjutkan sesi pembayaran Midtrans Snap
      security:
        - OrderAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSnapPaymentRequest'}
      responses:
        '201':
          description: Sesi pembayaran dibuat.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SnapPaymentSession'}
        '200':
          description: Sesi pembayaran yang masih aktif digunakan kembali.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SnapPaymentSession'}
        '401': {$ref: '#/components/responses/AccessTokenRequired'}
        '409': {$ref: '#/components/responses/Conflict'}
        '502':
          description: Penyedia pembayaran belum dapat membuat sesi.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '503':
          description: Pembuatan transaksi baru sedang dihentikan.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
  /api/v1/payments/midtrans/notification:
    post:
      operationId: midtransNotification
      tags: [Pembayaran]
      summary: Terima notifikasi status pembayaran Midtrans
      description: >-
        Endpoint memverifikasi signature_key dan nominal sebelum mengubah status order atau menerbitkan tiket.
        Notifikasi yang sama boleh dikirim berulang atau bersamaan; order tetap PAID dengan satu set e-ticket,
        dan email tiket tetap satu pekerjaan antrean. Pengiriman email terjadi di worker terpisah, sehingga
        kegagalan SMTP tidak memengaruhi respons webhook, status pembayaran, atau tiket.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MidtransNotification'}
      responses:
        '200': {description: Notifikasi diterima atau replay yang sudah diproses.}
        '400': {$ref: '#/components/responses/InvalidIDOrJSON'}
        '401': {description: Signature tidak valid.}
  /api/v1/orders/{orderID}/simulate-payment:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      operationId: simulatePayment
      tags: [Pembayaran]
      summary: Simulasikan pembayaran order
      description: Endpoint hanya didaftarkan saat APP_ENV=development. Pembayaran sukses menerbitkan snapshot e-ticket; replay sukses mengembalikan pembayaran dan tiket yang sama.
      security:
        - OrderAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulatePaymentRequest'
            examples:
              failed:
                value: {method: QRIS, result: FAILED}
              succeeded:
                value: {method: QRIS, result: SUCCEEDED}
      responses:
        '201':
          description: Pembayaran baru dicatat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResult'
        '200':
          description: Replay pembayaran atau percobaan ulang pembayaran gagal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResult'
        '400':
          $ref: '#/components/responses/InvalidIDOrJSON'
        '401':
          $ref: '#/components/responses/AccessTokenRequired'
        '404':
          description: Token salah atau order tidak ditemukan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                token:
                  value: {error: {code: NOT_FOUND, message: Data tidak ditemukan}}
                order:
                  value: {error: {code: ORDER_NOT_FOUND, message: Order tidak ditemukan}}
        '503':
          description: Pembuatan transaksi baru sedang dihentikan.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '409':
          description: Order kedaluwarsa/tidak dapat dibayar, atau pembayaran sukses tidak dapat diubah.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notPayable:
                  value: {error: {code: ORDER_NOT_PAYABLE, message: Order tidak dapat dibayar}}
                conflict:
                  value: {error: {code: PAYMENT_CONFLICT, message: Pembayaran berhasil dan tidak dapat diubah}}
        '410':
          $ref: '#/components/responses/AccessTokenExpired'
        '422':
          description: Metode atau hasil simulasi tidak valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: INVALID_REQUEST, message: Data pembayaran tidak valid}
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderID}/tickets:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    get:
      operationId: listOrderTickets
      tags: [E-ticket]
      summary: Daftar e-ticket order
      description: Daftar kosong sebelum pembayaran sukses.
      security:
        - OrderAccessToken: []
      responses:
        '200':
          description: Daftar snapshot e-ticket yang sudah diterbitkan.
          content:
            application/json:
              schema:
                type: object
                required: [tickets]
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/InvalidID'
        '401':
          $ref: '#/components/responses/AccessTokenRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/AccessTokenExpired'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    OrderAccessToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token opaque order dari respons checkout; terikat pada order dan berlaku sampai tengah malam WIB setelah tanggal konser.
    StaffSessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token sesi petugas opaque dari login; terpisah dari token akses order dan berlaku delapan jam.
...
  schemas:
...
    Event:
      type: object
      required: [id, artist, city, venue, address, startsAt, genre, status, publicationStatus, image, description, lineup, price, zones, ticketTiers, scheduleLocked]
      properties:
        id: {type: string}
        artist: {type: string}
        city: {type: string}
        venue: {type: string}
        address: {type: string}
        startsAt: {type: string, format: date-time}
        genre: {type: string}
        status: {type: string, enum: [Early Bird, Presale, Sold Out]}
        publicationStatus: {type: string, enum: [DRAFT, PUBLISHED, ARCHIVED]}
        image: {type: string}
        description: {type: string}
        lineup:
          type: array
          items: {type: string}
        price: {type: integer, minimum: 0, description: Harga minimum tier dalam rupiah.}
        zones:
          type: array
          items:
            $ref: '#/components/schemas/Zone'
        ticketTiers:
          type: array
          items:
            type: object
            required: [id, name, zoneId, price, availableQuantity, maxPerOrder, benefit, gate, seating]
            properties:
              id: {type: string}
              name: {type: string}
              zoneId: {type: string}
              price: {type: integer, minimum: 0, description: Harga satu tiket dalam rupiah.}
              availableQuantity: {type: integer, minimum: 0}
              maxPerOrder: {type: integer, minimum: 1}
              benefit: {type: string}
              gate: {type: string}
              seating: {type: string, enum: [assigned, free-standing]}
        scheduleLocked: {type: boolean, description: Jadwal tidak dapat diubah setelah order pertama dibuat.}
...